+++ Models/CallbackData/ConverstationStarted.cs
@@ -1,0 +1,7 @@
+﻿namespace Models.CallbackData
+{
+    public class ConverstationStarted : BaseIncomingMessage
+    {
+        public bool Subscribed { get; set; }
+    }
+}
