+++ Models/CallbackData/IncomingMessage.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-
-namespace ViberBotWebApp.Models.CallbackData
+﻿namespace Models.CallbackData
 {
     public class IncomingMessage : BaseIncomingMessage
     {
