+++ Models/DbModels/MatchResult.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace ViberBotWebApp.Models.DbModels
+﻿namespace Models.DbModels
 {
     public class MatchResult
     {
