+++ Models/Enums/State.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace ViberBotWebApp.Enums
+﻿namespace Models.Enums
 {
     public enum State
     {
