+++ Models/GetUserIdMessage.cs
@@ -1,0 +1,7 @@
+﻿namespace Models
+{
+    public class GetUserIdMessage
+    {
+        public string id { get; set; }
+    }
+}
