+++ Models/Location.cs
@@ -1,0 +1,8 @@
+﻿namespace Models
+{
+    public class Location
+    {
+        public double Lat { get; set; }
+        public double Lon { get; set; }
+    }
+}
