+++ Models/Member.cs
@@ -1,9 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace ViberBotWebApp.Models
+﻿
+namespace Models
 {
     public class Member
     {
