+++ Models/Sender.cs
@@ -1,0 +1,6 @@
+﻿namespace Models
+{
+    public class Sender : User
+    {
+    }
+}
