+++ Models/WebHookModel.cs
@@ -1,9 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
-namespace ViberBotWebApp.Models
+namespace Models
 {
     public class WebHookModel
     {
