+++ Models/WelcomeMessage.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace ViberBotWebApp.Models
+﻿namespace Models
 {
     public class WelcomeMessage
     {
