+++ Utils/HelperActions.cs
@@ -1,10 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Threading.Tasks;
 
-namespace ViberBotWebApp.Helpers
+namespace Utils
 {
     public static class HelperActions
     {
