+++ ViberBotWrapper/ActionsProvider/RequestProvider.cs
@@ -1,11 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net.Http;
 using System.Text.Json;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
-using ViberBotWebApp.Models;
+using Models;
 using static System.Net.Mime.MediaTypeNames;
 
 namespace ViberBotWebApp.ActionsProvider
