+++ ViberBotWrapper/ActionsProvider/StateManagerService.cs
@@ -1,10 +1,8 @@
-﻿using System;
+﻿using Models;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
-using ViberBotWebApp.Enums;
-using ViberBotWebApp.Models;
+using Models.Enums;
 
 namespace ViberBotWebApp.ActionsProvider
 {
