+++ ViberBotWrapper/DAL/DatabaseController.cs
@@ -2,14 +2,13 @@
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
-using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
+using Models;
+using Models.DbModels;
+using Utils;
 using ViberBotWebApp.DAL.DbConnection;
-using ViberBotWebApp.Helpers;
-using ViberBotWebApp.Models;
-using ViberBotWebApp.Models.DbModels;
 
 namespace ViberBotWebApp.DAL
 {
