--- ViberBotWrapper/Models/CallbackData/ConverstationStarted.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace ViberBotWebApp.Models.CallbackData
-{
-    public class ConverstationStarted : BaseIncomingMessage
-    {
-        public bool Subscribed { get; set; }
-    }
-}
