--- ViberBotWrapper/Models/GetUserIdMessage.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace ViberBotWebApp.Models
-{
-    public class GetUserIdMessage
-    {
-        public string id { get; set; }
-    }
-}
