--- ViberBotWrapper/Models/Location.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace ViberBotWebApp.Models
-{
-    public class Location
-    {
-        public double Lat { get; set; }
-        public double Lon { get; set; }
-    }
-}
